+++ history.md
@@ -3,6 +3,7 @@
 * Upgrade `multipart-post` dependency.
 * Replace `rainbow` dependency by `ansi`, which can also format columns. #55
 * Bug fix: `translation.yml` file wasn’t created on new projects.
+* New: Gracefully quit on interrupt.
 
 ## Version 1.8.0.1 / 2011-06-01